chore: improve Stepper cypress tests
